prandom: Add prandom() with xorshift128
This should give us a pretty fast PRNG with a longish period, portable
to other platforms without changing determinism. If we used the system's
rand() function from the C library, we would not be guaranteed that the
random numbers generated with the SDL build would match those generated
from the DOS build.