hashmap: Switch to linear probing hash map
Switch from a separate chaining hash map to a linear probing hash map.
Given that we use a 64-bit hash now, it's pretty unlikely to run into
collisions of our full hash value. We may still run into collisions at
the index (not hash) level, so we handle those in linear probing fashion
(find next empty slot; hash is hint to search in the array at a location
where we are likely to find the key).