hashmap: Use precomputed hashes
Using precomputed hashes allows us to re-add entries (necessary in
implementing delete) without having to remember the original full-length
key. It also allows us to avoid hashing multiple times for the same key
when checking to see if something exists before adding it. We
can also support using compile-time computed hashes in future.

Note this changes semantics of the hashtable somewhat. We no longer can
detect exact identical hashes as belonging to different keys, as the key
isn't available to the hashmap anymore. As such, we've updated the tests
for collisions to be bucket collisions rather than hash collisions.