test: Enable warnings, coverage, sanitizers
Enable the option of warnings, coverage, and sanitizers for each test
suite. This helps us catch errors with missing adding tests to the test
list, if those tests are static, for example; without warnings, we don't
catch that sort of thing. We also can benefit from getting more complete
coverage data when we ensure we build each test suite with coverage;
without this, we may, where we don't link with the guikit library, not
rack up coverage at all from that test suite.