sdl2: Render to texture
This enables easy scaling via GPU. SDL_BlitScaled() doesn't work with
paletted surfaces, and isn't hardware-accelerated anyway. Each frame,
convert our internal surface into a texture, and then use
SDL_RenderCopy() to copy-and-resize the texture to match the window
size. Because we are using SDL_RenderCopy(), the resize operation is
done on the GPU where possible.

An alternative would be to convert our internal paletted surface to a
32bpp surface, and then use SDL_BlitScaled() to go from there to the
window's surface; this wouldn't take advantage of any available hardware
scaling, however.