sdl2: Don't disable compositing
It's very annoying when the screen flickers (at least in the KDE Plasma
environment) when starting any of our demo applications. This happens
because SDL, by default, requests compositing disabled (ostensibly to
increase the framerate of the game or app). Since we don't run
fullscreen, setting this doesn't really make sense: the flickering would
be noticed and the rest of the desktop would get ugly. Request SDL not
to disable compositing, so that our window (and all others) remain
pretty and we avoid flickering screens.