tests: Expose the test runner as a library
This saves a tiny bit of compilation time as each test doesn't need to
recompile test.c. The largest benefit, however, is in packaging up the
test runner and its test.h header in a form convenient to link against:
as a CMake library. This allows projects depending on the GUI Kit to
also borrow its test framework for themselves, if they want to.